Replace old SprootDB with KnexDB
