+++ server/src/program.ts
@@ -7,7 +7,7 @@ import * as winston from "winston";
 
 import * as Constants from "@sproot/sproot-common/dist/utility/Constants";
 import { ISprootDB } from "@sproot/sproot-common/dist/database/ISprootDB";
-import { KnexDB } from "./database/KnexDB";
+import { SprootDB } from "./database/SprootDB";
 import { SDBUser } from "@sproot/sproot-common/dist/database/SDBUser";
 import { SensorList } from "./sensors/list/SensorList";
 import { OutputList } from "./outputs/list/OutputList";
@@ -25,7 +25,7 @@ export default async function setupAsync(): Promise<Express> {
   const knexConnection = await getKnexConnectionAsync();
   app.set("knexConnection", knexConnection);
 
-  const sprootDB = new KnexDB(knexConnection);
+  const sprootDB = new SprootDB(knexConnection);
   app.set("sprootDB", sprootDB);
   app.set("logger", logger);
